requirements.txt
acme==2.9.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argcomplete==2.0.0
attrs==23.1.0
Authlib==1.2.1
Babel==2.12.1
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bodhi-client==8.0.0
bottle==0.12.25
Brlapi==0.8.5
bs4==0.0.2
cached-property==1.5.2
cargo2rpm==0.1.15
CCColUtils==1.5
certbot==2.9.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.3
ConfigArgParse==1.7
configobj==5.0.8
construct==2.10.68
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dbus-next==0.2.3
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
dnf==4.19.0
docker==6.1.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
evdev==1.7.0
fake-useragent==1.5.0
fb-re2==1.0.7
fedora-review==0.10.0
fedora-third-party==0.10
fedpkg==1.44
feedparser==6.0.11
file-magic==0.4.0
filelock==3.12.2
future==0.18.3
gitdb==4.0.10
GitPython==3.1.40
Glances==3.4.0
googlesearch-python==1.2.3
gpg==1.20.0
gssapi==1.7.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
humanize==3.13.1
i3ipc==2.2.1
idna==3.4
importlib_metadata==7.0.2
Jinja2==3.1.3
josepy==1.13.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kerberos==1.3.0
keyboard==0.13.5
kitchen==1.2.6
koji==1.33.1
lameenc==1.7.0
lark==1.1.7
lark-parser==1.1.7
libcomps==0.1.20
libdnf==0.73.0
lockfile==0.12.2
lxml==4.9.3
MarkupSafe==2.1.3
mercurial==6.5.3
munch==2.5.0
ndg-httpsclient==0.5.1
nftables==0.1
numpy==1.26.4
openai==1.12.0
openidc-client==0.7.0
packaging==23.1
paramiko==3.4.0
parse==1.20.1
parsedatetime==2.6
pexpect==4.8.0
platformdirs==3.9.1
ply==3.11
podman-compose==1.0.6
powerline-status==2.8.3
progressbar2==3.53.2
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.1
PyAudio==0.2.14
pybeam==0.7
pycairo==1.25.1
pycparser==2.20
pycups==2.0.1
pycurl==7.45.2
pydantic==2.6.2
pydantic_core==2.16.3
pydub==0.25.1
pyee==11.1.0
pyenchant==3.2.2
pygame==2.5.2
pygit2==1.14.0
PyGObject==3.46.0
pykickstart==3.48
PyNaCl==1.5.0
pynput==1.7.6
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyppeteer==2.0.0
pyquery==2.0.0
pyRFC3339==1.1
pyroute2==0.7.3
pyrsistent==0.19.3
PySocks==1.7.1
python-augeas==1.1.0
python-bugzilla==3.2.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-fedora==1.1.1
python-gettext==4.0
python-manatools==0.0.4
python-utils==3.3.3
python-xlib==0.33
python-yubico==1.3.3
pytz==2024.1
pyudev==0.24.1
pyusb==1.2.1
pyxdg==0.27
PyYAML==6.0.1
referencing==0.30.2
requests==2.28.2
requests-gssapi==1.2.3
requests-html==0.10.0
requests-kerberos==0.12.0
requests-toolbelt==1.0.0
rpds-py==0.17.1
rpkg==1.66
rpm==4.19.1.1
rpmautospec==0.6.1
rpmautospec-core==0.1.4
rpmlint==2.5.0
rust2rpm==25.0.4
scipy==1.12.0
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sos==4.6.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.1
straight.plugin==1.5.0
systemd-python==235
templated-dictionary==1.4
tenacity==8.2.3
termcolor==2.3.0
texttable==1.6.4
timer==0.2.2
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tqdm==4.66.2
typing_extensions==4.9.0
ujson==5.8.0
urlgrabber==4.1.0
urllib3==1.26.18
virtualenv==20.21.1
w3lib==2.1.2
webrtcvad==2.0.10
websocket-client==1.3.3
websockets==10.4
whisper==1.1.10
zipp==3.17.0
zombie-imp==0.0.2
zope.component==5.0.1
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.5.2
zstandard==0.21.0
zstd==1.5.5.1
